auto-classification seems working. commit it immediately. It works with commit 6002178ecdcae21cf0536184af74e904ae7089f8 in module_tfn and commit 29b57d83b76afa342d9259e6e747da52e44a88de in module_visit

+++ projects/ospray/callback.cpp
@@ -67,7 +67,7 @@ void RenderWindow(GLFWwindow *window)
        [&](const std::vector<float> &c, const std::vector<float> &a) {
          std::vector<float> o(a.size() / 2);
          for (size_t i = 0; i < a.size() / 2; ++i) { o[i] = a[2 * i + 1]; }
-         volume.GetTransferFunction().Update(c.data(), o.data(), c.size() / 3, 1, o.size(), 1);
+         volume.GetTransferFunction().Update(c.data(), o.data(), (int)c.size() / 3, 1, (int)o.size(), 1);
          ClearOSPRay();
        });
 #endif
